feat(electronics): fold the workspace toolbar into a Circuit palette tab
Replaces the floating bottom "Schematic | Components | PCB | View | Finish" bar
with a single "Circuit" tab in the main borland ToolPalette. Its tool row is
context-aware by the schematic/board view: schematic shows select/move/wire/
label/delete + the component palette; board shows select/move/route/length-tune/
delete/auto-route/place-unplaced + layer & grid controls. When no circuit is
open the tab offers the entry point (Edit Circuit, or New PCB Board).

The Circuit tab is coupled to the electronics workspace: selecting it enters the
circuit (or shows New-PCB), and selecting any other tab exits back to the
mechanical workspace; entering via the panel / double-click pins the tab too.

- ToolbarTab gains "circuit" (+ colors/theme/description, getAllTabs entry).
- CircuitTabTools + symbol-icons extracted from the deleted ElectronicsToolbar.
- ToolPalette: renderTabContent special-cases circuit; handleTabClick +
  autoSwitchTab own the enter/exit coupling.
